Remove raw_ prefix
The prefix reads well in isolation but obscures the  membership to
the "family" of `scenario_demo` datasets. A consistent prefix plays
well with autocompletion, both in rstudio and the terminal. e.g.
ls data-raw/scenario*

Also the prefix "raw" doesn't add much meaning -- every data in
data-raw may be raw.