New scenario demo 2020 (#38)
Closes #37  

* Remove raw_ prefix
The prefix reads well in isolation but obscures the  membership to
the "family" of `scenario_demo` datasets. A consistent prefix plays
well with autocompletion, both in rstudio and the terminal. e.g.
ls data-raw/scenario*
Also the prefix "raw" doesn't add much meaning -- every data in
data-raw may be raw.

* Use r2dii.scenario 0.0.0.9002
* add data dictionary entry for scenaro_demo_2020
* remove legacy scenario_demo documentation
* update data_dictionary
* document scenario_demo_2020
* format scenario_demo_2020
* update data_dictionmary ref
* Minor rewording and styling
* Update spelling wordslist
* use_tidy_desription()
* Use add_market_share_columns from .scenario .9002
  ```
  # Install the relevant version of r2dii.scenario with:
  remotes::install_github(
    "2degreesinvesting/r2dii.scenario@0.0.0.9002"
  )
  ```

Co-authored-by: Jackson Hoffart <jackson.hoffart@gmail.com>