Assume a single raw dataset region_isos
This commit simplifies the file structure and code to read a
single raw region_isos dataset. This does the job for now so I
prefer to keep it simple.

But later we should need more complex code to store and read
multiple raw datasets (I think one each year). When that happens
we can revert this commit.

Also the new name of the raw data to more closely follows the
naming style of other datasets. The consistent prefix makes it
easier to see that region_isos.R and region_isos_* are related.