Side effects
This commit adds the side effects of the previous one. These result
from sourcing data-raw/, document()-ing, and updating regression
references.

It will most likely result in merge conflicts that are difficult to
resolve. You may want to drop this commit and instead reproduce it
doing yourself the steps I described in the previous paragraph.