Warn that ald_demo is superseded
ald_demo no longer is an object. We still support the symbol
`ald_demo` by allowing users to use it and assigning the
contents of `abcd_demo` on the fly -- with a warning that
`ald_demo` is superseded by `abcd_demo`.

As far as I know `delayAssign()` is the only way to inject
"function-like" behaviour (like throwing a warning) when the user
calls a non-function symbol.